+++ index.js
@@ -3,4 +3,4 @@ module.exports.Batch = require('./lib/batch');
 module.exports.File  = require('./lib/file');
 module.exports.Utils  = require('./lib/utils');
 module.exports.Validate  = require('./lib/validate');
-module.exports.isBankHoliday  = require('./lib/bank-holidays/index').isBankHoliday;
+module.exports.isBankHoliday  = require('./lib/bank-holidays/utils').isBankHoliday;
